Error handling and functionality
Added a new variable to the Book class, expanded the GetBookByTitle function to also operate by title and it has been renamed accordingly.
Error handling as been added to AddBookToLibrary so that books with the same title but different authors can exist, but books which have identical authors and titles cannot.