Added functionality
Created a main loop and some input handling so that the user can run functions.
Added a Loaner class so that loaners can loan books from the library.
Added a DateTime object variable to the Book and Library classes to that loan times can be tracked.
Added a function for displaying all books in the console, called "DisplayAllBooks."
Added a WIP function called "LoanBook" used for loaning a book from the library.
Added a function for retrieving loan time from a book.